chore(refactor): move btrfs code out of zfs package
